requirements.txt
flask
requests
statistics
python-dotenv
flask-caching